+++ CHANGELOG.md
@@ -38,6 +38,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Wrong taxon name relationship soft validation message for genera
 - Modal confirmation its not displaying in manage synonyms section [#1627]
 - Manage synonyms includes combinations [#1628]
+- Recent and per params in source filter and controller
 
 [#446]: https://github.com/SpeciesFileGroup/taxonworks/issues/446
 [#856]: https://github.com/SpeciesFileGroup/taxonworks/issues/856